adding tmpfilecache default folder, improving tmpfile cache output and api

+++ utils/tmpfilecache.js
@@ -4,16 +4,33 @@ var async = require('async');
 var config = require('../config');
 var sha256 = require('./sha256');
 
+var cache_miss = function(cache, key, callback){
+    this.cache = cache;
+    this.key = key;
+    this.file = cache.key2File(this.key);
+    this.callback = callback;
+};
+cache_miss.prototype.set_file_data = function(data){
+    this.cache.put(this.key, data, this.callback);
+};
+cache_miss.prototype.retry_get = function(){
+    this.cache.get(this.key, this.callback);
+};
+
+
 var cache = {
     hash_key : function(key){
         var match = /^(.*?)(\..*)?$/.exec(key);
         return sha256(match[1]) + (match[2] || '');
     },
+    key2File : function(key){
+        var root = path.resolve(config("tmpfilecache").path);
+        return path.join(root, this.hash_key(key));
+    },
     checkValidity : function(file, callback){
         fs.stat(file, function(err, stats){
             if (err) { callback(err); return ; }
             var now = (new Date()).getTime();
-            console.log(stats.atime.getTime(), " + ", config("tmpfilecache").maxObjectLifetime, " >= " , now, " ? ", stats.atime.getTime() + config("tmpfilecache").maxObjectLifetime >= now);
             if (stats.atime.getTime() + config("tmpfilecache").maxObjectLifetime >= now ) { 
                 callback(null, {
                     path : file,
@@ -25,26 +42,17 @@ var cache = {
         })
     },
     get :  function(key, callback){
-        var root = path.resolve(config("tmpfilecache").path);
-        var file = path.join(root, this.hash_key(key));
-        
-        cache.checkValidity(file, function(err, stats){
+        cache.checkValidity(cache.key2File(key), function(err, stats){
             callback(err, stats);
         });
     },
     put : function(key, data, callback){
-        var root = path.resolve(config("tmpfilecache").path);
-        var file = path.join(root, this.hash_key(key));
-        
-        fs.writeFile(file, data, callback);
-        
+        fs.writeFile(cache.key2File(key), data, callback);
     },
     fetch : function(key, oncachemiss, callback){
         this.get(key, function(err, data){
             if(!data || err){
-                oncachemiss(function(data){
-                    cache.put(key, data, callback);
-                });
+                oncachemiss(new cache_miss(cache, key, callback));
             } else {
                 callback(null, data);
             }
@@ -53,25 +61,30 @@ var cache = {
     
     cleanupTimeout : 0,
     cleanup : function(){
+        console.log('Cleaning up tmpcache.');
         var root = path.resolve(config("tmpfilecache").path);
         var setCleanupTimeout = function(){
             if (cache.cleanupTimeout) {
                 return;
             }
+            var delay = config("tmpfilecache").cleanupInterval;
+            console.log('Running tmpcache cleanup in : ', (delay/1000.0), ' seconds.');
             cache.cleanupTimeout = setTimeout(function(){
                 cache.cleanupTimeout = 0;
                 cache.cleanup();
-            }, config("tmpfilecache").cleanupInterval);
+            }, delay);
         };
         
         fs.readdir(root, function(err, files){
             async.each(files, function(subfile, callback){
+                if (subfile == '.gitignore') {
+                    return;
+                }
                 var file = path.join(root, subfile);
                 cache.checkValidity(file, function (err, filestats){
                     if(!filestats) {
                         fs.unlink(file);
-                        console.log('cleanup tmpcache file : ', file, err, filestats);
-                        
+                        console.log('   tmpcache file removed : ', file);
                     }
                 });
             }, function(err){
@@ -80,7 +93,5 @@ var cache = {
         })
     }
 };
-
-cache.cleanup();
 
 module.exports = cache;
